Move comps to own folder
