+++ src/pages/admin/components/TeamDetails.tsx
